requirements.txt
pydantic>=2.0
openenv-core>=0.2.0
fastapi
uvicorn
numpy
pandas
openai
torch
matplotlib
unsloth
git+https://github.com/arcee-ai/mergekit.git
accelerate
transformers
datasets